[YP-356] refactor: 작가 검색 페이지 컴포넌트 경로 수정
