+++ src/components/creators/CreatorsSearchView/components/Banner/defines/styles.ts
