+++ src/components/creators/CreatorsSearchView/components/CreatorsView/CreatorsFilter/defines/constants.ts
@@ -1,7 +1,7 @@
 import {
 	CreatorsKeywordType,
 	SortByType,
-} from "@/components/creators/components/CreatorsView/components/CreatorsFilter/defines/type";
+} from "@/components/creators/CreatorsSearchView/components/CreatorsView/CreatorsFilter/defines/type";
 
 export const CATEGORY_SELECT_BOX_WIDTH = 120;
 export const BUDGET_SELECT_BOX_WIDTH = 100;
