+++ src/components/creators/CreatorsSearchView/defines/styles.ts
